Changed OpenGL ES 1.1 to 2.0. Delete Mac, Qt.

+++ classes/mog/core/TouchEventListener.cpp
@@ -1,6 +1,5 @@
 #include "mog/core/TouchEventListener.h"
 #include "mog/core/Engine.h"
-#include "mog/base/Entity.h"
 
 using namespace mog;
 
@@ -17,7 +16,7 @@ TouchEventListener::~TouchEventListener() {
 void TouchEventListener::touchBegin(const Touch &touch, const shared_ptr<Entity> &entity) {
     if (!this->enabled) return;
     
-    unsigned long long tid = ((uintptr_t)entity.get() << 32) + touch.touchId;
+    unsigned long long tid = ((unsigned long long)entity.get() << 32) + touch.touchId;
     
     if (this->onTouchEnterEvent) {
         this->onTouchEnterEvent(touch, entity);
@@ -37,7 +36,7 @@ void TouchEventListener::touchBegin(const Touch &touch, const shared_ptr<Entity>
 void TouchEventListener::touchMove(const Touch &touch, const shared_ptr<Entity> &entity) {
     if (!this->enabled) return;
     
-    unsigned long long tid = ((uintptr_t)entity.get() << 32) + touch.touchId;
+    unsigned long long tid = ((unsigned long long)entity.get() << 32) + touch.touchId;
     
     if (this->isTouchBegan(tid)) {
         if (this->onTouchDragEvent) {
@@ -69,7 +68,7 @@ void TouchEventListener::touchMove(const Touch &touch, const shared_ptr<Entity> 
 void TouchEventListener::touchEnd(const Touch &touch, const shared_ptr<Entity> &entity) {
     if (!this->enabled) return;
     
-    unsigned long long tid = ((uintptr_t)entity.get() << 32) + touch.touchId;
+    unsigned long long tid = ((unsigned long long)entity.get() << 32) + touch.touchId;
     
     if (this->isTouchBegan(tid)) {
         if (this->onTouchEndEvent) {
